pyproject.toml
[project]
name = "sam"
version = "0.1.0"
description = "SAM: Script's Advanced Moderation (features)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "discord-py>=2.6.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.25",
]
